+++ src/components/layout/public.ts
@@ -1,3 +1,6 @@
-export * from "./app-sidebar"
+export * from "./sidebar"
+export * from "./workspace-header"
 export * from "./breadcrumbs"
 export * from "./section"
+export * from "./sidebar-context"
+export * from "./workspace-layout"